169 bug (#173)
* refactor: 팝업 뷰컨트롤러 폴더 세분화

* fix: 버튼이 없을 경우 팝업의 높이가 맞게 조절되도록 수정

* fix: WebSocket이 안전하게 종료될 수 있도록 수정

* refactor: 채팅 뷰컨트롤러 코드 리팩토링

* feat: 알림 삭제 구현

* fix: 요청 거절 응답 구조체 수정 및 수락/거절 버튼 오류 수정

* fix: currentPage 추가, OtherImageData 구조체에 값 추가

* fix: 토큰 플러그인 refresh path 변경

* feat: 요청 클릭 시, 상세뷰로 넘어가게 설정

* fix: 이미지에 scaleAspectFill 적용

* fix: 후기 별점을 소수점 한자리 수까지만 보내도록 수정

* fix: 응답 구조체 수정

* fix: 모든 아이돌을 선택된 아이돌로 보내던 오류 수정

* feat: 자기소개가 없는 경우 안내문구 추가

* fix: 소켓매니저에 토큰 담기, 연결 유지 중일때만 다음 메세지를 대기하도록 추가

* fix: 메세지 뷰의 조건 변경

* fix: secret 제거

* feat: 탈퇴한 회원이었을 경우 뷰 다르게 보이도록

* feat: 다른 사람 프로필-게시물 조회에서 게시물 상세로 이동하도록

* feat: 게시물상세-채팅버튼에서 탈퇴/채팅방이 이미 있는 경우 고려

* fix: 알림 뷰 네비게이션바 조건 추가

* fix: 채팅뷰 띄울 때 chatRoomId만 필요하도록 수정

+++ Duckmelang/Duckmelang/Network/Responses/ReviewResponse.swift
@@ -21,7 +21,7 @@ public struct ReviewInformation: Codable {
 // 후기글 작성 성공 후 받아오는 구조체
 public struct PostReviewResponse: Codable {
     public let reviewId: Int
-    public let score: Int
+    public let score: Double
     public let content: String
     public let receiverId: Int
 }
